+++ assets/javascripts/kitten/components/form/text-input-with-button/index.js
@@ -61,6 +61,7 @@ export class TextInputWithButton extends PureComponent {
     error: PropTypes.bool,
     disabled: PropTypes.bool,
     tiny: PropTypes.bool,
+    huge: PropTypes.bool,
     modifier: PropTypes.string,
     inputValue: PropTypes.string,
     // DEPRECATED: do not use prop `value`. Use `buttonValue` instead.
@@ -73,6 +74,7 @@ export class TextInputWithButton extends PureComponent {
     error: false,
     disabled: false,
     tiny: false,
+    huge: false,
     buttonValue: 'Button',
     modifier: 'beryllium',
   }
@@ -83,6 +85,7 @@ export class TextInputWithButton extends PureComponent {
       error,
       disabled,
       tiny,
+      huge,
       value,
       buttonValue,
       modifier,
@@ -99,6 +102,7 @@ export class TextInputWithButton extends PureComponent {
           error={error}
           disabled={disabled}
           tiny={tiny}
+          huge={huge}
           value={inputValue}
         />
         <StyledButton