+++ assets/javascripts/kitten/components/form/text-input-with-button/test.js
@@ -91,6 +91,16 @@ describe('<TextInputWithButton />', () => {
     })
   })
 
+  describe('with huge prop', () => {
+    beforeEach(() => {
+      component = renderer.create(<TextInputWithButton huge />).toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
   describe('with buttonProps prop', () => {
     beforeEach(() => {
       component = renderer